// src/pages/CompanyPages/settings/screens/LoyalityProgramSection/hooks/types.ts
export interface FormProps {
  levels?: any;
  requirements?: any;
  max_percent?: string | number;
  give_cashback_after?: string | number;
  base_level?: any;
  base_name?: any;
  base_percent?: any;
  useProgram?: any;
  usePoint?: any;
}

export interface IEmpty {
  empty?: boolean;
  type?: string;
}